move and rename rrweb types file
