+++ packages/replay/src/types/rrweb.ts
